+++ docs/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 2026-04-13
+### Changed
+- Permission dependencies now support optional location-scoped checks. Endpoints accept optional `location_id` query param which is forwarded to permission resolver.
+
 ## 2026-03-16
 ### Added
 - repo operating system docs